+++ zson/zson.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/brimsec/zq/compiler/ast"
 	"github.com/brimsec/zq/zng"
-	"github.com/brimsec/zq/zng/resolver"
 )
 
 // Implied returns true for primitive types whose type can be inferred
@@ -57,7 +56,7 @@ func SelfDescribing(typ zng.Type) bool {
 	return false
 }
 
-func LookupType(zctx *resolver.Context, zson string) (zng.Type, error) {
+func LookupType(zctx *TypeContext, zson string) (zng.Type, error) {
 	zp, err := NewParser(strings.NewReader(zson))
 	if err != nil {
 		return nil, err
@@ -69,6 +68,6 @@ func LookupType(zctx *resolver.Context, zson string) (zng.Type, error) {
 	return NewAnalyzer().convertType(zctx, ast)
 }
 
-func TranslateType(zctx *resolver.Context, astType ast.Type) (zng.Type, error) {
+func TranslateType(zctx *TypeContext, astType ast.Type) (zng.Type, error) {
 	return NewAnalyzer().convertType(zctx, astType)
 }
